Chore: 코드리뷰 반영하여 변수명 변경, 라우터 분리 등
